requirements.txt
bottle
google-cloud-storage
redis